Add replica info to the set data structure
Add a new `repinfo` field (removing `this`) that contains the
information about the current replica, other participating replicas and
all participating replicas. This infomation could have been derieved
from the previous structure but having dedicated fields makes it more
explicit.

We could have also removed all the replica information from the data
structure and let it be part of the replica state (outside of the set)
which I think is a better solution. Eventually I feel that we'll get
there.