Increment replica timestamps on deletes
We increment the timestamps on delete operations at the source replica
and for all delete downstream operations received at a downstream
replica we add the timestamp of the delete operation to the interval
version vector corresponding to the source replica at the downstream
replica.

NOTE: This change deviates from the original paper where timestamps are
only updated for delete operations.